+++ ui/src/main/java/io/snabble/sdk/ui/cart/shoppingcart/utils/TextFieldManager.kt
@@ -16,6 +16,10 @@ internal class TextFieldManager(
         focusManager.clearFocus()
         keyboardController?.hide()
     }
+    
+    fun showKeyboard(){
+        keyboardController?.show()
+    }
 
     fun moveFocusToNext() {
         focusManager.moveFocus(FocusDirection.Next)