+++ ui/src/main/java/io/snabble/sdk/ui/remotetheme/SnabblePrimaryButton.kt
@@ -1,0 +1,28 @@
+package io.snabble.sdk.ui.remotetheme
+
+import android.content.Context
+import android.util.AttributeSet
+import com.google.android.material.button.MaterialButton
+import io.snabble.sdk.Snabble
+import io.snabble.sdk.ui.R
+
+/**
+ * A default Materialbutton which automatically sets the remote theme colors of the
+ * current checked in project.
+ */
+class SnabblePrimaryButton @JvmOverloads constructor(
+    context: Context,
+    attrs: AttributeSet? = null,
+    defStyleAttr: Int = R.attr.materialButtonStyle,
+) : MaterialButton(context, attrs, defStyleAttr) {
+
+    init {
+        setProjectAppTheme()
+    }
+
+    private fun setProjectAppTheme() {
+        val project = Snabble.checkedInProject.value
+        setBackgroundColor(context.getPrimaryColorForProject(project))
+        setTextColor(context.getOnPrimaryColorForProject(project))
+    }
+}